+++ src/common/init.c
@@ -144,7 +144,7 @@ static void destroyState(void)
     ffPlatformDestroy(&instance.state.platform);
     yyjson_doc_free(instance.state.configDoc);
     yyjson_mut_doc_free(instance.state.resultDoc);
-    ffStrbufDestroy(&instance.state.migrateConfigPath);
+    ffStrbufDestroy(&instance.state.genConfigPath);
 }
 
 void ffDestroyInstance(void)